No longer install `/etc/ulogd.conf`.
We can't install a default `ulogd.conf` because most `ulogd` packages do that already. Rather, we install an example `ulogd.conf` in `/usr/share/pcaping-strong-iptables`.

This commit also adds `all` targets to all the `Makefile` files so that a default run of `make` will not attempt to install the package.